+++ workshop1/eca3_jit_and_scan.py
@@ -36,7 +36,7 @@ def main(
     print("initialising state...")
     match init:
         case "middle":
-            state = jnp.zeros(width, dtype=int)
+            state = jnp.zeros(width, dtype=jnp.uint8)
             state = state.at[width//2].set(1)
         case "random":
             key = jax.random.key(seed)
@@ -46,7 +46,7 @@ def main(
                 minval=0,
                 maxval=2, # not included
                 shape=(width,),
-                dtype=int,
+                dtype=jnp.uint8,
             )
     print("initial state:", state)
 
@@ -83,13 +83,13 @@ def main(
         
 def simulate(
     rule: int,
-    init_state: jax.Array,    # int[width]
+    init_state: jax.Array,    # uint8[width]
     height: int,
-) -> jax.Array:               # int[height, width]
+) -> jax.Array:               # uint8[height, width]
     # parse rule
     rule_uint8 = jnp.uint8(rule)
     rule_bits = jnp.unpackbits(rule_uint8, bitorder='little')
-    rule_table = rule_bits.reshape(2,2,2).astype(int)
+    rule_table = rule_bits.reshape(2,2,2)
 
     # pad initial state
     init_state = jnp.pad(init_state, 1, mode='wrap')
